Guts all moved to nitf module
