+++ six/modules/c++/six/include/six/ByteProvider.h
@@ -23,8 +23,7 @@
 #ifndef __SIX_BYTE_PROVIDER_H__
 #define __SIX_BYTE_PROVIDER_H__
 
-#include <nitf/ImageBlocker.hpp>
-#include <nitf/NITFBufferList.hpp>
+#include <nitf/ByteProvider.hpp>
 #include <mem/SharedPtr.h>
 #include <six/Container.h>
 #include <six/NITFWriteControl.h>
@@ -48,100 +47,12 @@ namespace six
  * rather than one contiguous block of memory in order to minimize the number of
  * copies.
  */
-class ByteProvider
+class ByteProvider : public nitf::ByteProvider
 {
-public:
-    /*!
-     * \return The total number of bytes in the NITF
-     */
-    nitf::Off getFileNumBytes() const
-    {
-        return mFileNumBytes;
-    }
-
-    /*!
-     * Given a range of rows from [startRow, startRow + numRows), provide the
-     * number of bytes that will appear in the NITF on disk (including NITF
-     * file header, image subheader(s), and DES subheader(s) and data).  Calling
-     * this method repeatedly, eventually providing the entire range of the
-     * image, will produce the total number of bytes in the full NITF.
-     *
-     * \param startRow The global start row in pixels as to where these pixels
-     * are in the image.  If this is a multi-segment NITF, this is still simply
-     * the global pixel location.
-     * \param numRows The number of rows.
-     *
-     * \return The associated number of bytes in the NITF
-     */
-    nitf::Off getNumBytes(size_t startRow, size_t numRows) const;
-
-    /*!
-     * The caller provides an AOI of the pixel data.  This method provides back
-     * a list of contiguous buffers corresponding to the raw NITF bytes for
-     * this portion of the file.  If this AOI is in the middle of an image
-     * segment, this will be simply a buffer list of length 1 consisting of the
-     * input pointer (no copy occurs).  Otherwise, pointers to various headers
-     * (file header, image subheader(s), DES subheader and data) will
-     * be in the buffer list before, in the middle, and/or after the image
-     * data.  If this method is called multiple times with AOIs that
-     * eventually consist of the entire image, and the raw bytes are written
-     * out to disk in order with respect to the start pixel rows this method is
-     * called with (or out of order but seeking to the provided file offset
-     * each time), and in the order contained in the buffer list, it will form
-     * a valid NITF.
-     *
-     * \note This method does not perform byte swapping on the pixel data for
-     * efficiency reasons, but NITFs are written out in big endian order.  This
-     * means that on a little endian system, you must byte swap the pixel data
-     * prior to calling this method.
-     *
-     * \note This method does not perform any blocking on the pixel data.  If
-     * the NITF is blocked, the pixel data must already be in a contiguous
-     * blocked memory layout, including pad pixels for partial blocks.
-     *
-     * \param imageData The image data pixels to write.  The underlying type
-     * will be complex short or complex float based on the complex data sent
-     * in during initialize.  Must be in big endian order and blocked if the
-     * NITF is.
-     * \param startRow The global start row in pixels as to where these pixels
-     * are in the image.  If this is a multi-segment NITF, this is still simply
-     * the global pixel location.
-     * \param numRows The number of rows in the provided 'imageData'
-     * \param[out] fileOffset The offset in bytes in the NITF where these
-     * buffers should be written
-     * \param[out] buffers One or more pointers to raw bytes of data.  These
-     * should be written out in the order they are provided in the buffer list.
-     * The pointers point to the provided pixel data and, if required, one or
-     * more NITF headers.  No copies occur, so these buffers are only valid for
-     * the lifetime of this provider object (since this object owns the raw
-     * bytes for the NITF headers) and the lifetime of the passed-in image
-     * data.
-     */
-    void getBytes(const void* imageData,
-                  size_t startRow,
-                  size_t numRows,
-                  nitf::Off& fileOffset,
-                  nitf::NITFBufferList& buffers) const;
-
-    /*!
-     * \return ImageBlocker with settings in sync with how the image will be
-     * blocked in the NITF
-     */
-    std::auto_ptr<const nitf::ImageBlocker> getImageBlocker() const;
-
 protected:
     /*!
-     * Constructor
-     *
-     * \param schemPaths Directories or files of schema locations
-     * \param maxProductSize The max number of bytes in an image segment.
-     * By default this is set automatically for you based on NITF file rules.
-     */
-    ByteProvider();
-
-    /*!
      * Initialize the byte provider.  Must be called in the constructor of
-     * inheriting classes
+     * inheriting classes.
      *
      * \param container Container initialized with all associated data
      * \param xmlRegistry XML registry
@@ -161,7 +72,7 @@ protected:
 
     /*!
      * Initialize the byte provider.  Must be called in the constructor of
-     * inheriting classes
+     * inheriting classes.
      *
      * \param writer Writer.  Must be initialized via
      * NITFWriteControl::initialize() and have all desired product size and
@@ -170,26 +81,6 @@ protected:
      */
     void initialize(const NITFWriteControl& writer,
                     const std::vector<std::string>& schemaPaths);
-
-private:
-    void checkBlocking(size_t seg,
-                       size_t startGlobalRowToWrite,
-                       size_t numRowsToWrite) const;
-
-private:
-    size_t mNumCols;
-    size_t mOverallNumRowsPerBlock;
-    std::vector<size_t> mNumRowsPerBlock;
-    size_t mNumColsPerBlock;
-    size_t mNumBytesPerRow;
-
-    std::vector<sys::byte> mFileHeader;
-    std::vector<std::vector<sys::byte> > mImageSubheaders;
-    std::vector<sys::byte> mDesSubheaderAndData;
-    std::vector<nitf::Off> mImageSubheaderFileOffsets;
-    std::vector<NITFSegmentInfo> mImageSegmentInfo;
-    nitf::Off mDesSubheaderFileOffset;
-    nitf::Off mFileNumBytes;
 };
 }
