+++ six/modules/c++/six/source/ByteProvider.cpp
@@ -20,22 +20,12 @@
  *
  */
 
-#include <limits>
-#include <sstream>
-
-#include <six/NITFWriteControl.h>
+#include <str/Convert.h>
+#include <logging/NullLogger.h>
 #include <six/ByteProvider.h>
 
 namespace six
 {
-ByteProvider::ByteProvider() :
-    mNumCols(0),
-    mOverallNumRowsPerBlock(0),
-    mNumColsPerBlock(0),
-    mNumBytesPerRow(0)
-{
-}
-
 void ByteProvider::initialize(mem::SharedPtr<Container> container,
                               const XMLControlRegistry& xmlRegistry,
                               const std::vector<std::string>& schemaPaths,
@@ -86,263 +76,71 @@ void ByteProvider::initialize(mem::SharedPtr<Container> container,
 void ByteProvider::initialize(const NITFWriteControl& writer,
                               const std::vector<std::string>& schemaPaths)
 {
-    // Now we can get the file headers and offsets we want
-    writer.getFileLayout(schemaPaths,
-                         mFileHeader,
-                         mImageSubheaders,
-                         mDesSubheaderAndData,
-                         mImageSubheaderFileOffsets,
-                         mImageSegmentInfo,
-                         mDesSubheaderFileOffset,
-                         mFileNumBytes);
+    // Sanity check the container
+    mem::SharedPtr<const Container> container = writer.getContainer();
 
-    // getFileLayout() is going to ensure this is one vertically stacked image
-    // and not 2+ unrelated images with different sizes, so this is a reliable
-    // way to get the # of cols and bytes/pixel
-    const six::Data* const data = writer.getContainer()->getData(0);
-    mNumCols = data->getNumCols();
+    if (container->getNumData() == 0)
+    {
+        throw except::Exception(Ctxt(
+                "Write control must be initialized first"));
+    }
 
+    // We currently do not support the case where there are 2+ unrelated SIDDs
+    // in a file (but one logical SIDD which spans multiple image segments
+    // and/or contains SICD XML is ok).  This is a limitation in how
+    // nitf::ByteProvider computes row offsets.  So, ensure this constraint is
+    // met.
+    bool haveDerived(false);
+    for (size_t ii = 0; ii < container->getNumData(); ++ii)
+    {
+        if (container->getData(ii)->getDataType() == DataType::DERIVED)
+        {
+            if (!haveDerived)
+            {
+                haveDerived = true;
+            }
+            else
+            {
+                throw except::Exception(Ctxt(
+                        "Don't currently support more than one SIDD image"));
+            }
+        }
+    }
+
+    // Create XML strings
+    // This memory must stay around until the call to the
+    // base class's initialize() method
+    logging::NullLogger logger;
+    std::vector<std::string> xmlStrings(container->getNumData());
+    std::vector<PtrAndLength> desData(xmlStrings.size());
+    for (size_t ii = 0; ii < xmlStrings.size(); ++ii)
+    {
+        std::string& xmlString(xmlStrings[ii]);
+        xmlString = six::toValidXMLString(container->getData(ii),
+                                          schemaPaths,
+                                          &logger,
+                                          writer.getXMLControlRegistry());
+        desData[ii].first = xmlString.c_str();
+        desData[ii].second = xmlString.length();
+    }
+
+    // Get blocking info
     const Options& options(writer.getOptions());
+    const Parameter zero(0);
 
-    if (options.hasParameter(six::NITFWriteControl::OPT_NUM_ROWS_PER_BLOCK))
-    {
-        mOverallNumRowsPerBlock = options.getParameter(
-                six::NITFWriteControl::OPT_NUM_ROWS_PER_BLOCK);
-    }
+    const size_t numRowsPerBlock = static_cast<sys::Uint32_T>(
+            options.getParameter(NITFWriteControl::OPT_NUM_ROWS_PER_BLOCK,
+                                 zero));
 
-    size_t numColsWithPad;
-    if (options.hasParameter(six::NITFWriteControl::OPT_NUM_COLS_PER_BLOCK))
-    {
-        mNumColsPerBlock = options.getParameter(
-                six::NITFWriteControl::OPT_NUM_COLS_PER_BLOCK);
+    const size_t numColsPerBlock = static_cast<sys::Uint32_T>(
+            options.getParameter(NITFWriteControl::OPT_NUM_COLS_PER_BLOCK,
+                                 zero));
 
-        numColsWithPad = nitf::ImageSubheader::getActualImageDim(
-                mNumCols, mNumColsPerBlock);
-    }
-    else
-    {
-        numColsWithPad = mNumCols;
-    }
-
-    mNumBytesPerRow = numColsWithPad * data->getNumBytesPerPixel();
-    if (mOverallNumRowsPerBlock == 0)
-    {
-        mNumRowsPerBlock.resize(mImageSegmentInfo.size());
-        std::fill(mNumRowsPerBlock.begin(), mNumRowsPerBlock.end(), 0);
-    }
-    else
-    {
-        mNumRowsPerBlock = getImageBlocker()->getNumRowsPerBlock();
-    }
-}
-
-void ByteProvider::checkBlocking(size_t seg,
-                                 size_t startGlobalRowToWrite,
-                                 size_t numRowsToWrite) const
-{
-    if (mOverallNumRowsPerBlock != 0)
-    {
-        const NITFSegmentInfo& imageSegmentInfo(mImageSegmentInfo[seg]);
-        const size_t segStartRow = imageSegmentInfo.firstRow;
-
-        // Need to start on a block boundary
-        const size_t segStartRowToWrite =
-                startGlobalRowToWrite - segStartRow;
-        const size_t numRowsPerBlock(mNumRowsPerBlock[seg]);
-        if (segStartRowToWrite % numRowsPerBlock != 0)
-        {
-            std::ostringstream ostr;
-            ostr << "Trying to write starting at segment " << seg
-                 << "'s row " << segStartRowToWrite
-                 << " which is not a multiple of " << numRowsPerBlock;
-            throw except::Exception(Ctxt(ostr.str()));
-        }
-
-        // Need to end on a block boundary or the end of the segment
-        if (numRowsToWrite % numRowsPerBlock != 0 &&
-            segStartRowToWrite + numRowsToWrite < imageSegmentInfo.numRows)
-        {
-            std::ostringstream ostr;
-            ostr << "Trying to write " << numRowsToWrite
-                 << " rows at global start row " << startGlobalRowToWrite
-                 << " starting at segment " << seg
-                 << " which is not a multiple of " << numRowsPerBlock
-                 << " (segment has start = " << imageSegmentInfo.firstRow
-                 << ", num = " << imageSegmentInfo.numRows << ")";
-            throw except::Exception(Ctxt(ostr.str()));
-        }
-    }
-}
-
-nitf::Off ByteProvider::getNumBytes(size_t startRow, size_t numRows) const
-{
-    nitf::Off numBytes(0);
-
-    const size_t imageDataEndRow = startRow + numRows;
-
-    for (size_t seg = 0; seg < mImageSegmentInfo.size(); ++seg)
-    {
-        // See if we're in this segment
-        const NITFSegmentInfo& imageSegmentInfo(mImageSegmentInfo[seg]);
-
-        size_t startGlobalRowToWrite;
-        size_t numRowsToWrite;
-        if (imageSegmentInfo.isInRange(startRow, numRows,
-                                       startGlobalRowToWrite,
-                                       numRowsToWrite))
-        {
-            checkBlocking(seg, startGlobalRowToWrite, numRowsToWrite);
-            const size_t segStartRow = imageSegmentInfo.firstRow;
-
-            if (startRow <= segStartRow)
-            {
-                // We have the first row of this image segment, so we're
-                // responsible for the image subheader
-                if (seg == 0)
-                {
-                    // For the very first image segment, we're responsible for
-                    // the file header too
-                    numBytes += mFileHeader.size();
-                }
-
-                numBytes += mImageSubheaders[seg].size();
-            }
-
-            const size_t numRowsPerBlock(mNumRowsPerBlock[seg]);
-            if (numRowsPerBlock != 0 &&
-                imageDataEndRow >= imageSegmentInfo.endRow())
-            {
-                const size_t numLeftovers =
-                        imageSegmentInfo.numRows % numRowsPerBlock;
-                if (numLeftovers != 0)
-                {
-                    // We need to finish the block
-                    const size_t numPadRows = numRowsPerBlock - numLeftovers;
-                    numRowsToWrite += numPadRows;
-                }
-            }
-
-            numBytes += numRowsToWrite * mNumBytesPerRow;
-
-            if (seg == mImageSegmentInfo.size() - 1 &&
-                imageSegmentInfo.endRow() == imageDataEndRow)
-            {
-                // When we write out the last row of the last image segment, we
-                // tack on the DES(s)
-                numBytes += mDesSubheaderAndData.size();
-            }
-        }
-    }
-
-    return numBytes;
-}
-
-void ByteProvider::getBytes(const void* imageData,
-                            size_t startRow,
-                            size_t numRows,
-                            nitf::Off& fileOffset,
-                            nitf::NITFBufferList& buffers) const
-{
-    fileOffset = std::numeric_limits<nitf::Off>::max();
-    buffers.clear();
-
-    const size_t imageDataEndRow = startRow + numRows;
-
-    size_t numPadRowsSoFar(0);
-
-    for (size_t seg = 0; seg < mImageSegmentInfo.size(); ++seg)
-    {
-        // See if we're in this segment
-        const NITFSegmentInfo& imageSegmentInfo(mImageSegmentInfo[seg]);
-
-        size_t startGlobalRowToWrite;
-        size_t numRowsToWrite;
-        if (imageSegmentInfo.isInRange(startRow, numRows,
-                                       startGlobalRowToWrite,
-                                       numRowsToWrite))
-        {
-            checkBlocking(seg, startGlobalRowToWrite, numRowsToWrite);
-            const size_t segStartRow = imageSegmentInfo.firstRow;
-
-            if (startRow <= segStartRow)
-            {
-                // We have the first row of this image segment, so we're
-                // responsible for the image subheader
-                if (seg == 0)
-                {
-                    // For the very first image segment, we're responsible for
-                    // the file header too
-                    fileOffset = 0;
-                    buffers.pushBack(mFileHeader);
-                }
-
-                if (buffers.empty())
-                {
-                    fileOffset = mImageSubheaderFileOffsets[seg];
-                }
-                buffers.pushBack(mImageSubheaders[seg]);
-            }
-
-            // Figure out what offset of 'imageData' we're writing from
-            const size_t startLocalRowToWrite =
-                    startGlobalRowToWrite - startRow + numPadRowsSoFar;
-            const sys::byte* imageDataPtr =
-                    static_cast<const sys::byte*>(imageData) +
-                    startLocalRowToWrite * mNumBytesPerRow;
-
-            if (buffers.empty())
-            {
-                const size_t rowsInSegmentSkipped =
-                        startRow - segStartRow;
-
-                fileOffset = mImageSubheaderFileOffsets[seg] +
-                        mImageSubheaders[seg].size() +
-                        rowsInSegmentSkipped * mNumBytesPerRow;
-            }
-
-            const size_t numRowsPerBlock(mNumRowsPerBlock[seg]);
-            if (numRowsPerBlock != 0 &&
-                imageDataEndRow >= imageSegmentInfo.endRow())
-            {
-                const size_t numLeftovers =
-                        imageSegmentInfo.numRows % numRowsPerBlock;
-                if (numLeftovers != 0)
-                {
-                    // We need to finish the block
-                    // This is already accounted for in the incoming image
-                    const size_t numPadRows = numRowsPerBlock - numLeftovers;
-                    numRowsToWrite += numPadRows;
-                    numPadRowsSoFar += numPadRows;
-                }
-            }
-
-            buffers.pushBack(imageDataPtr, numRowsToWrite * mNumBytesPerRow);
-
-            if (seg == mImageSegmentInfo.size() - 1 &&
-                imageSegmentInfo.endRow() == imageDataEndRow)
-            {
-                // When we write out the last row of the last image segment, we
-                // tack on the DES(s)
-                buffers.pushBack(mDesSubheaderAndData);
-            }
-        }
-    }
-}
-
-std::auto_ptr<const nitf::ImageBlocker> ByteProvider::getImageBlocker() const
-{
-    std::vector<size_t> numRowsPerSegment(mImageSegmentInfo.size());
-    for (size_t ii = 0; ii < mImageSegmentInfo.size(); ++ii)
-    {
-        numRowsPerSegment[ii] = mImageSegmentInfo[ii].numRows;
-    }
-
-    std::auto_ptr<const nitf::ImageBlocker> blocker(new nitf::ImageBlocker(
-            numRowsPerSegment,
-            mNumCols,
-            mOverallNumRowsPerBlock,
-            mNumColsPerBlock));
-
-    return blocker;
+    // Do the full initialization
+    nitf::Record record = writer.getRecord();
+    nitf::ByteProvider::initialize(record,
+                                   desData,
+                                   numRowsPerBlock,
+                                   numColsPerBlock);
 }
 }
